requirements.txt
Django>=4.0
djangorestframework>=3.12
Pillow>=9.0